+++ View/Controllers/CategoryController.cs
@@ -1,6 +1,7 @@
 ﻿using DataAnimals.DTO.Animal;
 using DataAnimals.DTO.Category;
 using Microsoft.AspNetCore.Mvc;
+using System.Net.Http.Headers;
 using System.Net.Mime;
 using System.Text;
 using System.Text.Json;
@@ -18,8 +19,27 @@ namespace View.Controllers
         public async Task<IActionResult> Index([FromQuery] string? filterQuery, bool isAccess)
         {
             var client = _httpClientFactory.CreateClient();
-
             List<CategoryDto> category = new List<CategoryDto>();
+            var tokenJson = HttpContext.Session.GetString("Jwt");
+
+            // Kiểm tra xem token có tồn tại không
+            if (string.IsNullOrEmpty(tokenJson))
+            {
+                // Nếu không tồn tại, chuyển hướng đến trang đăng nhập
+                return RedirectToAction("PageLogin", "Account");
+            }
+
+            // Giải mã JSON để lấy giá trị của token
+            var tokenObj = JsonSerializer.Deserialize<Dictionary<string, string>>(tokenJson);
+            if (tokenObj == null || !tokenObj.ContainsKey("jwtToken"))
+            {
+                return RedirectToAction("PageLogin", "Account");
+            }
+            var token = tokenObj["jwtToken"];
+
+            // Thêm token vào header Authorization của yêu cầu HTTP
+            Console.WriteLine("JWT Token: " + token); // Ghi log token để kiểm tra
+            client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
             var data = await client.GetAsync($"https://localhost:7035/api/Category/GetAll?name={filterQuery}&isAccess={isAccess}");
             data.EnsureSuccessStatusCode();
             category.AddRange(await data.Content.ReadFromJsonAsync<IEnumerable<CategoryDto>>());
@@ -28,8 +48,29 @@ namespace View.Controllers
         //Add data
         public async Task<IActionResult> Create()
         {
+
             var client = _httpClientFactory.CreateClient();
             List<AnimalDto> animal = new List<AnimalDto>();
+            var tokenJson = HttpContext.Session.GetString("Jwt");
+
+            // Kiểm tra xem token có tồn tại không
+            if (string.IsNullOrEmpty(tokenJson))
+            {
+                // Nếu không tồn tại, chuyển hướng đến trang đăng nhập
+                return RedirectToAction("PageLogin", "Account");
+            }
+
+            // Giải mã JSON để lấy giá trị của token
+            var tokenObj = JsonSerializer.Deserialize<Dictionary<string, string>>(tokenJson);
+            if (tokenObj == null || !tokenObj.ContainsKey("jwtToken"))
+            {
+                return RedirectToAction("PageLogin", "Account");
+            }
+            var token = tokenObj["jwtToken"];
+
+            // Thêm token vào header Authorization của yêu cầu HTTP
+            Console.WriteLine("JWT Token: " + token); // Ghi log token để kiểm tra
+            client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
             var data = await client.GetAsync("https://localhost:7035/api/Animal/GetAll");
             data.EnsureSuccessStatusCode();
             animal.AddRange(await data.Content.ReadFromJsonAsync<IEnumerable<AnimalDto>>());
@@ -42,6 +83,26 @@ namespace View.Controllers
             try
             {
                 var client = _httpClientFactory.CreateClient();
+                var tokenJson = HttpContext.Session.GetString("Jwt");
+
+                // Kiểm tra xem token có tồn tại không
+                if (string.IsNullOrEmpty(tokenJson))
+                {
+                    // Nếu không tồn tại, chuyển hướng đến trang đăng nhập
+                    return RedirectToAction("PageLogin", "Account");
+                }
+
+                // Giải mã JSON để lấy giá trị của token
+                var tokenObj = JsonSerializer.Deserialize<Dictionary<string, string>>(tokenJson);
+                if (tokenObj == null || !tokenObj.ContainsKey("jwtToken"))
+                {
+                    return RedirectToAction("PageLogin", "Account");
+                }
+                var token = tokenObj["jwtToken"];
+
+                // Thêm token vào header Authorization của yêu cầu HTTP
+                Console.WriteLine("JWT Token: " + token); // Ghi log token để kiểm tra
+                client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
                 var httpReq = new HttpRequestMessage()
                 {
                     Method = HttpMethod.Post,
@@ -61,7 +122,7 @@ namespace View.Controllers
             }
             catch
             {
-                return View("Info");
+                return View("Warn");
             }
         }
         //Delete
@@ -71,12 +132,32 @@ namespace View.Controllers
             try
             {
                 var client = _httpClientFactory.CreateClient();
+                var tokenJson = HttpContext.Session.GetString("Jwt");
+
+                // Kiểm tra xem token có tồn tại không
+                if (string.IsNullOrEmpty(tokenJson))
+                {
+                    // Nếu không tồn tại, chuyển hướng đến trang đăng nhập
+                    return RedirectToAction("PageLogin", "Account");
+                }
+
+                // Giải mã JSON để lấy giá trị của token
+                var tokenObj = JsonSerializer.Deserialize<Dictionary<string, string>>(tokenJson);
+                if (tokenObj == null || !tokenObj.ContainsKey("jwtToken"))
+                {
+                    return RedirectToAction("PageLogin", "Account");
+                }
+                var token = tokenObj["jwtToken"];
+
+                // Thêm token vào header Authorization của yêu cầu HTTP
+                Console.WriteLine("JWT Token: " + token); // Ghi log token để kiểm tra
+                client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
                 var httpReponse = await client.DeleteAsync("https://localhost:7035/api/Category/Delete?id=" + id);
                 httpReponse.EnsureSuccessStatusCode();
             }
             catch
             {
-                return View("Info");
+                return View("Warn");
 
             }
             return RedirectToAction("Index", "Category");
@@ -85,7 +166,26 @@ namespace View.Controllers
         public async Task<IActionResult> UpdateCtx(int Id)
         {
             var client = _httpClientFactory.CreateClient();
-            //
+            var tokenJson = HttpContext.Session.GetString("Jwt");
+
+            // Kiểm tra xem token có tồn tại không
+            if (string.IsNullOrEmpty(tokenJson))
+            {
+                // Nếu không tồn tại, chuyển hướng đến trang đăng nhập
+                return RedirectToAction("PageLogin", "Account");
+            }
+
+            // Giải mã JSON để lấy giá trị của token
+            var tokenObj = JsonSerializer.Deserialize<Dictionary<string, string>>(tokenJson);
+            if (tokenObj == null || !tokenObj.ContainsKey("jwtToken"))
+            {
+                return RedirectToAction("PageLogin", "Account");
+            }
+            var token = tokenObj["jwtToken"];
+
+            // Thêm token vào header Authorization của yêu cầu HTTP
+            Console.WriteLine("JWT Token: " + token); // Ghi log token để kiểm tra
+            client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
             CategoryDto category = new CategoryDto();
             var data = await client.GetAsync("https://localhost:7035/api/Category/Get-by-ID?id=" + Id);
             data.EnsureSuccessStatusCode();
@@ -103,23 +203,50 @@ namespace View.Controllers
         [HttpPost]
         public async Task<IActionResult> UpdateCategory(AddCategoryDto category, int Id)
         {
+            try
+            {
+                var client = _httpClientFactory.CreateClient();
+                var tokenJson = HttpContext.Session.GetString("Jwt");
 
-            var client = _httpClientFactory.CreateClient();
-            var httpReq = new HttpRequestMessage()
-            {
-                Method = HttpMethod.Put,
-                RequestUri = new Uri("https://localhost:7035/api/Category/Update?id=" + Id),
-                Content = new StringContent(JsonSerializer.Serialize(category), Encoding.UTF8,
-                    MediaTypeNames.Application.Json)
-            };
-            var httpRes = await client.SendAsync(httpReq);
-            httpRes.EnsureSuccessStatusCode();
-            var data = await httpRes.Content.ReadFromJsonAsync<AddCategoryDto>();
-            if (data != null)
-            {
-                return RedirectToAction("Index", "Category");
+                // Kiểm tra xem token có tồn tại không
+                if (string.IsNullOrEmpty(tokenJson))
+                {
+                    // Nếu không tồn tại, chuyển hướng đến trang đăng nhập
+                    return RedirectToAction("PageLogin", "Account");
+                }
+
+                // Giải mã JSON để lấy giá trị của token
+                var tokenObj = JsonSerializer.Deserialize<Dictionary<string, string>>(tokenJson);
+                if (tokenObj == null || !tokenObj.ContainsKey("jwtToken"))
+                {
+                    return RedirectToAction("PageLogin", "Account");
+                }
+                var token = tokenObj["jwtToken"];
+
+                // Thêm token vào header Authorization của yêu cầu HTTP
+                Console.WriteLine("JWT Token: " + token); // Ghi log token để kiểm tra
+                client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
+                var httpReq = new HttpRequestMessage()
+                {
+                    Method = HttpMethod.Put,
+                    RequestUri = new Uri("https://localhost:7035/api/Category/Update?id=" + Id),
+                    Content = new StringContent(JsonSerializer.Serialize(category), Encoding.UTF8,
+                        MediaTypeNames.Application.Json)
+                };
+                var httpRes = await client.SendAsync(httpReq);
+                httpRes.EnsureSuccessStatusCode();
+                var data = await httpRes.Content.ReadFromJsonAsync<AddCategoryDto>();
+                if (data != null)
+                {
+                    return RedirectToAction("Index", "Category");
+                }
+                return View("Detail");
             }
-            return View("Detail");
+            catch
+            {
+                return View("Warn");
+            }
+            
         }
     }
 }
